Add new CSS and JS files and modify HTML structure
Four new CSS files have been added to the HTML structure for enhancing the UI. A new JavaScript file, 'jquery-3.7.1.min.js', is added for JQuery functionalities. Plus another JavaScript file for element labels. Changes are made in the 'nav' and 'search-area' classes for better alignment. Also, the label text for 'search' is set to lowercase.