Add tooltip functionality for elements with titles
A new JavaScript file, `element-label.js`, has been created to handle tooltip functionality for all elements with a 'title' attribute. It is designed to display the title of these elements as text within a 'label' <div> which appears when the mouse is hovered over the element. The positioning of this tooltip label is calculated on mouseover to ensure it doesn't exceed the boundaries of the window.