Update search area CSS styling
This change updates the CSS properties of the search area in the navigation bar. The width property of the search area has been adjusted to have a minimum and maximum limit. The update also rectifies the width of the floating input within the search area making the layout more flexible and adaptative.