Update element-label.js to improve tooltip functionality
To prevent the default browser tooltips from showing, the code was updated to instead store the tooltip content in a 'data-title' attribute. Then, a custom tooltip system was added by creating a div with class "label", to which the 'data-title' attribute is assigned. The 'mouseover' event was also updated to prevent it from defaulting and to grab the text from the 'data-title' attribute instead.