Add "greet" command to Tauri application
A new greeting function, "greet", has been added to the Tauri application. This command accepts a name as an input and returns a string greeting. Incorporating this greeting feature enhances the user interaction experience with the application. The invoke_handler for the greet function has also been integrated into the main function.