Refactor label creation process in element-label.js
The creation of a new div with class 'label' was moved outside of the 'loadLabels' function, making it a global scope event. Earlier, it was only created if there were no 'label' elements in the document body, and this logic was nested inside the 'loadLabels' function. The restructuring ensures the label div is immediately available, regardless of call frequency to 'loadLabels'.