Refactor database list and implement voice search
The database list has been made more efficient by improving the processes of loading, filtering, and pagination. The voice search capability has also been added as this would be a more convenient way for users to search the database. This is achieved through the use of an external library, Voice.js.