Add DirectoryList class in directory-list.js
A new class, DirectoryList, has been added in directory-list.js. This class is used to manage and manipulate a list, including loading view with given queries, fetching list items from an API endpoint, constructing pagination for the list, and building a HTML representation of the list items. It also performs various actions, such as hiding certain elements and loading labels.