Add search function and Filemaker import form
Included a new search function to refine the items within the database list. Also built a form for importing databases from Filemaker. Adjustments were made to ensure compatibility of the updated url scheme. Additionally, condition for list building was amended to facilitate display of the new import form when item list is empty.