Refactor directory list javascript file
The loadView method is now being called when the '#hero button' is clicked rather than at initial runtime. Also, the URL for fetching locations from the API has been updated to use a dynamic base URL. A new method 'search' has been added to load a view based on a query and return the items. The 'buildListHTML' method has been refactored to accept items as argument and the list item creation includes a new, clickable area.