Implement localStorage for loaded database
A loaded database is now stored and retrieved from the window's localStorage. This was achieved by changing logic flow for the display of the edit button, adding listeners for loadExternalView, unloadExternalView, and a search event. The event handlers also make use of the localStorage to store or remove the loadedDatabase.