Refactor voice-search.js to simplify keypress and interim events
Changed the 'keypress' event in voice-search.js to directly call the 'search' function, removing the need for an anonymous function. Also, the 'interim' event handler was simplified by calling 'search' directly with the transcript, instead of setting the search input value first. A new 'end' event handler was also added to stop the voice input when finished.