Refactor database options manager in JS
This update refactors the `database-options-manager.js` in several ways. A significant change is the removal of 'id' as a global variable and instead passing it as a parameter. Furthermore, the 'save' function has been adjusted to make ajax calls to update options & columns. Also, unnecessary variables and comments have been cleaned up to improve code's readability.