Add resetListElement function in page-navigation.js
A new function, resetListElement, was added to the page-navigation.js file to reset the layout of the list element. This function is called at various points in the code - when loading an external view, editing a view, unloading an external view, searching, and clicking the edit button. This modification ensures the list element consistently resets its layout based on certain user activities or interactions.