Update Filemaker data import process
This code revises the process of importing data from Filemaker. Changes include the update of the Filemaker-API path and the extraction of the database building process to a new function buildView. It also adds error messaging in the filemaker credentials form and logic to filter databases based on a new search event. Furthermore, the data import process has been restructured to provide better user feedback during the operation, including estimates of completion time and improved error handling.