Add "OK" button and callbacks to alert popup
The alert popup function has been updated to accept an additional optional callback function, which is executed when the "OK" button is clicked. The popup itself has been modified to include a "Cancel" or a "Close" button depending on whether an "OK" function is provided or not. These changes increase the versatility of the function and allow for additional user interaction opportunities.