Implement localStorage and form submission in inventorying
The code has been updated to store the selected item in localStorage for better user experience. A form submission event has been implemented which conditions on whether the item is new or is being updated. When adding or updating items, all the relevant columns are iterated to fetch values to be sent to the server. If the item is not found, an alert is issued to the user. Additionally, the code includes some small fixes like correcting the label for on an input field.