Add export functionality to page-navigation.js
An "Export" button was introduced to the page-navigation.js file to provide users with the ability to export data. The display settings for this button were also adjusted in various situations, such that it appears and hides appropriately. Additionally, the 'resetListElement' has been relocated within the "search" document event for better performance.