Add retail and mp values to print function in database-list.js
The JavaScript file for handling the database list has been modified to include two values, retail and mp, in the print function. This allows pertinent product information to be printed along with other attributes of the item. These values are initialized and populated only if they exist in the item's attributes.