Update Authentication API and improve user experience
This commit updates the reference to the latest version of the Authentication API. Additionally, it augments the visibility of error messages in the login form and tweaks the behavior after logging in. Specifically, the "loginPrompt" is removed from local storage and non-admin users are prevented from accessing certain resources. It also cleans up some unnecessary commented out code.