Update authentication API and add cookie expiration
This commit updates the version of the Authentication API being used, and modifies the authentication logic to include cookie expiration. If the site is running in debug mode (http), the cookie will now expire in 365 days; otherwise, it will expire when the browser is closed, enhancing security for users in the production environment.