Add ETA calculation to FileMaker data import
The update introduces an ETA (Estimated Time of Arrival) calculation to the FileMaker data import function. The ETA is dynamically calculated based on the average items processed per second. The calculated value is then formatted and included in the loading message displayed to the user during the data importing process. This change makes it clear to the user how long the importing process is likely to take, improving the user experience.