Add history record handling to history.js
Extended the functionality of the history.js module to now include detailed handling of history records. The additions include the definition of the ActionType object and associated functions to transform these types into various formats. Also added are functions to retrieve and add history records from/to the server, and the ability to parse these records from received JSON format.