Convert ajax data to JSON format in location.js
Previously, in the AJAX call made in location.js, the data was being passed directly as an object. This commit changes it so that the data is first converted into a JSON string before it's sent. The use of JSON.stringify helps ensure that the server will correctly interpret the received data.