Add batch record addition functionality
The commit introduces a new function, batchAddRecords, to handle the addition of multiple records at a specified location in the database. It includes time tracking elements such as a countdown and estimated time of arrival (ETA) for completion. It also refactors the code to normalize potential ID and date field conflicts in the incoming data.