Enhance CSV handling and form building functionality
The CSV data parsing was improved and the regular expression for data sanitisation was updated to allow apostrophes. Two key functions, `buildMardensPriceForm` and `buildPrintSection`, were overhauled. In the former, code was added to handle the title and unique column items based on the `column-attribute-change` event. In the latter, the loading and display of print options were implemented, including hiding and showing relevant input fields based on selected category. The use of template literals was introduced in alert statements to correctly display error messages.