Updated visibility and structure of form elements in database-options-form
This commit made changes to the visibility of the retail and Mardens price toggles: they're no longer hidden. Additionally, the structure of discount categories was revised with an added ID for improved indexing and identification. Introduced default discount value to the first category. The per-category delete button code seems to be removed.