Refactored "mardens-price" logic and changed print implementation
The logic for handling "mardens-price" option was refactored to cover edge cases and improve code readability. Also, a try-catch block was added for error handling. The way data is sent to the print window was also modified. Instead of using $.get to fetch the data and then writing it into the print window, the URL is directly opened within the print window for simplicity.